+++ library/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to the library will be documented in this file.
 
+## vX.X.X (Month DD, YYYY)
+
+- Rename `NoPipe` type to `SchemaWithoutPipe`
+
 ## v0.41.0 (September 01, 2024)
 
 - Change `reference` property of all action base types to be less strict (issue #799)